moved directories, removed an unecessary step in finishing tween.js to improve framerate spikes
