+++ src/index.js
@@ -1,6 +1,6 @@
 "use strict";
 
 import "./index.html";
-import { Scroller } from "./components/scroller.js";
+import { Scroller } from "./components/scroller/scroller.js";
 
 Scroller.draw();
